# Use Accelerate framework on Apple silicon