TL;DR: x2 performance improvement in the Encoder on Apple silicon

Modern Apple products come with a special AMX matrix coprocessor. Currently, the only way to utilise it in your code is through [APPLE's Accelerate framework](https://developer.apple.com/documentation/accelerate). Among many things, it provides a [BLAS implementation](https://developer.apple.com/documentation/accelerate/blas) that apparently runs on the AMX.

Using the AMX for Whisper model inference is extremely beneficial for the Encoder part of the model. The tensor sizes in the encoder are big enough to compensate for the Accelerate framework overhead and even bring about x2 better performance compared to my previous F16 implementation. It is a shame not to use it, so this PR adds support for this.

Here is before and after comparison of running the `large` model on `jfk.wav`:

```java
# without Accelerate
whisper_print_timings:     load time =  1018.34 ms
whisper_print_timings:      mel time =    13.85 ms
whisper_print_timings:   sample time =     2.11 ms
whisper_print_timings:   encode time =  9123.14 ms / 285.10 ms per layer
whisper_print_timings:   decode time =  1154.06 ms / 36.06 ms per layer
whisper_print_timings:    total time = 11311.87 ms

# with Accelerate
whisper_print_timings:     load time =  1002.72 ms
whisper_print_timings:      mel time =    13.43 ms
whisper_print_timings:   sample time =     2.90 ms
whisper_print_timings:   encode time =  4071.11 ms / 127.22 ms per layer
whisper_print_timings:   decode time =  1117.53 ms / 34.92 ms per layer
whisper_print_timings:    total time =  6208.27 ms
```

This finally gives me the answer how PyTorch's implementation of the encoder was outperforming `whisper.cpp` encoder on my MacBook. I am pretty sure that PyTorch also ends up using the Accelerate framework under the hood. With this change, the encoder performance between the two implementations is now comparable.

My tests show that it is not beneficial to use CBLAS for the decoder part of the model. Multi-threaded Neon SIMD is significantly faster. Likely due to the smaller tensor sizes.